Separated component printing sizes. Added Automatic expire date input.

+++ flask_app/kit_routes.py
@@ -28,6 +28,9 @@ def kit_delete(kit_id):
 	if not current_user.logged_in():
 		return redirect(url_for('login'))
 	kit = Kit.query.get(kit_id)
+	current_time = datetime.today()
+	if (current_time - kit.date_entered).total_seconds() > 3*3600:
+		return redirect(url_for('kit', kit_id=kit_id))
 	db.session.delete(kit)
 	db.session.commit()
 	return redirect(url_for('kits'))
@@ -83,8 +86,8 @@ def add_kit_redirect():
 
 	kit_label_size = request.form.get('kit_label_size')
 	kit_label = int(request.form.get('kit_label'))
-	comp_label_size = request.form.get('comp_label_size')
-	comp_label = int(request.form.get('comp_label'))
+	comp_label_s = int(request.form.get('comp_label_s'))
+	comp_label_m = int(request.form.get('comp_label_m'))
 
 	batchpartnum = 1
 	while batchpartnum <= kit_label:
@@ -94,9 +97,14 @@ def add_kit_redirect():
 
 	for name, comp_num, part_num, lot_num, condition in zip(names, comp_nums, comp_part_nums, comp_lot_nums, conditions):
 		batchpartnum = 1
-		while batchpartnum <= comp_label:
+		while batchpartnum <= comp_label_s:
 			printcont = (name, request.form.get("exp_date"), datetime.now())
-			print_label(printcont, comp_label_size, str(batchpartnum) + '/' + str(comp_label))
+			print_label(printcont, 's', str(batchpartnum) + '/' + str(comp_label_s))
+			batchpartnum += 1
+
+		while batchpartnum <= comp_label_m:
+			printcont = (name, request.form.get("exp_date"), datetime.now())
+			print_label(printcont, 'm', str(batchpartnum) + '/' + str(comp_label_m))
 			batchpartnum += 1
 
 		component = Component(
